Force end FTM session
The original FTM code indicate that the FTM session will be closed only
when it got response. However, when the traffic become heavier, the FTM
packets might be dropped due to collision or other reasons. It leads to
the FTM session will never be closed.

Workaround to force ending the FTM sessions when the simulation is over.